move current home page to ghofund/v01
